backend/requirements.txt
# FastAPI va Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Autentifikatsiya va Xavfsizlik
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Ma'lumotlar validatsiyasi
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment
python-dotenv==1.0.0

# WebSocket
websockets==12.0

# Sana va vaqt
python-dateutil==2.8.2

# Fayl yuklash
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# PDF va Excel generatsiya
reportlab==4.0.7
openpyxl==3.1.2

# QR Code
qrcode==7.4.2
pillow==10.1.0

# Printer (ESC/POS)
python-escpos==2.2.0
pyusb==1.2.1